check local hub from js client before contacting proxy hub
Now when messaging the hub, the javascript client first gets the
signalling image from the localhost hub, and only proceeds to
send an XML-RPC call to the remote proxy hub if that image
retrieval succeeds.  Otherwise it assumes the localhost hub
is dead so there's no point attempting SAMP activity.

While this will slow things down a little bit, since it needs
to wait for a localhost HTTP round trip before initiating the
remote host communications, it's not going to be by much
(since the localhost hub is much closer) and it means that
useless messages don't get queued up on the proxy hub if there's
no likelihood of a localhost hub coming to get them.