improve security info displayed to user on registration
The registration popup for TLS profile registration now contains
the Referer header and the proxy hub location alongside the
application name.  Referer might not be present or the proxy hub
might be lying about it, but the proxy hub should be reliable,
especially if it's HTTPS (a warning is issued if not).
I think these are basically the only bits of information that
the localhost hub can find at all reliably.