nomenclature: refer to "relay" not "bouncer" or "proxy hub"
Use the single term "relay" or "hub relay" rather than talking
about bouncing or a proxy hub.  This is more compact and less
confusing, especially since it makes it looks like there is
a hub on a remote host.  The "localhost hub" can now in
most cases be called just the "hub".

This change of nomenclature changes the public protocol as well
as class names, variable names and documentation, but there is
no change to the TLS SAMP architecture or functionality.