adjust credential presentation
Changed JSAMP, replacing AuthHeaderControl with CredentialPresenter.
This is JSAMP 9a25c20cf279.

This enables me to pass the referer header (and possibly others in
future) in the register() method's identity-info argument, rather
than having to shove referer into the SampCall object.
I think that's the right thing to do, though now I think about it
I'm not 100% sure it's an improvement.  For some reason it took me
a vast amount of time to implement this change.

In any case the credential presentation to users is now somewhat
more configurable.