include sampjs examples in .war file
Fix the makefile to include the content of sampjs/examples.
The point of this is so that the same examples can be seen
from the tomcat servlet in both its HTTP and HTTPS forms,
so I can point at them both for testing purposes from
https://wiki.ivoa.net/twiki/bin/view/IVOA/WebSampHttps

This is super-lazy - the content doesn't really belong in
tlsamp, but that's where I've got a tomcat application and
it's the easiest way to make the content available from
HTTP and HTTPS simultaneously.